app_setup/requirements.txt
psycopg2-binary==2.9.1
Flask==2.0.2
Flask_RESTful==0.3.9